# optimisation of grayscale morphology operators