This PR has been made following the issue #597  : optimization of grayscale morphology operators

it aims to work on the internal implementation of the `Mask` and associated functions to optimize the execution time, as well as add parallel versions of some public functions using rayon.

in the first commit that i have made, i have added bench for the `Mask` creation methods and have reworked the implementation of the `from_image` methods, which allowed a 4x speed increase (200 ms to 50ms) for a large image (200 by 200 pixels). i would expect the improvement to be less significant for smaller images, although i have not tested it.

it was brought to my attention that i was accessing image pixels in col-major order. That was due to a misunderstanding of the image layout from my part. In light of this, i have reworked the implementation of the `Mask` to facilitate and incentivize access in row-major order by storing the positions row by row.

i have also been advise to change (i16, i16) to Point<i16> for readability, which i will do in a further commit.

Finally, through my testing, i tried to use `unsafe_get_pixel` to optimize my `get_pixel` calls. While it was very efficient, i found that accessing the buffer directly and using the `chunks` method was faster by around 17% (~65 ms to ~50ms). i tested it several times, and the slowest `chunks` time was faster than the fastest `unsafe_get_pixel` by 10 ms. 

i believe that it might be interesting to look at for other uses of `unsafe_get_pixel` which traverse the image line by line, as it potentially could be the case that directly traversing a `[u8]` might allow the compiler to do better optimizations than when these accesses are obfuscated through `unsafe_get_pixel`