@Morgane55440 2024-05-08
next i'm changing the `(i16, i16)` to `Point`, benching `grayscale_erode` and `grayscale_dilate`, and testing different implementations