@Morgane55440 2024-05-08
> ```rust
> let elements = image
>             .enumerate_pixels()
>             .filter(|(_, _, &p)| p[0] != 0)
>             .map(|(x, y, _)| (x as i16 - center_x as i16, y as i16 - center_y as i16))
>             .collect();
> Self { elements }
> ```

thanks a ton ! it is just as fast, if not slightly faster (seems maybe 5-10 ms faster on average, but there was overlap)