@Morgane55440 2024-05-08
i am not super knowledgeable on all the ìmage` methods, but i'll look into them more