@cospectrum 2024-05-08
The new release is scheduled for May 14th, if anything. I think you will make it in time