@Morgane55440 2024-05-09
initial stae of benches : 
```
test morphology::benches::bench_grayscale_op_dilate_big_image_diamond                 ... bench:  86,049,620 ns/iter (+/- 13,300,790)
test morphology::benches::bench_grayscale_op_dilate_big_image_point                   ... bench:   2,315,520 ns/iter (+/- 188,327)
test morphology::benches::bench_grayscale_op_dilate_medium_image_diamond              ... bench:   3,373,130 ns/iter (+/- 193,688)
test morphology::benches::bench_grayscale_op_dilate_medium_image_large_square         ... bench: 131,351,560 ns/iter (+/- 8,071,737)
test morphology::benches::bench_grayscale_op_dilate_medium_image_point                ... bench:      99,243 ns/iter (+/- 13,569)
test morphology::benches::bench_grayscale_op_dilate_small_image_diamond               ... bench:     208,150 ns/iter (+/- 14,842)
test morphology::benches::bench_grayscale_op_dilate_small_image_large_square          ... bench:   7,243,630 ns/iter (+/- 1,650,297)
test morphology::benches::bench_grayscale_op_dilate_small_image_point                 ... bench:       5,766 ns/iter (+/- 178)
test morphology::benches::bench_grayscale_op_erode_big_image_diamond                  ... bench:  88,131,170 ns/iter (+/- 9,589,572)
test morphology::benches::bench_grayscale_op_erode_big_image_point                    ... bench:   2,311,410 ns/iter (+/- 726,459)
test morphology::benches::bench_grayscale_op_erode_medium_image_diamond               ... bench:   3,503,005 ns/iter (+/- 314,480)
test morphology::benches::bench_grayscale_op_erode_medium_image_large_square          ... bench: 130,097,690 ns/iter (+/- 6,730,746)
test morphology::benches::bench_grayscale_op_erode_medium_image_point                 ... bench:      82,883 ns/iter (+/- 5,617)
test morphology::benches::bench_grayscale_op_erode_small_image_diamond                ... bench:     212,093 ns/iter (+/- 40,138)
test morphology::benches::bench_grayscale_op_erode_small_image_large_square           ... bench:   7,109,105 ns/iter (+/- 483,316)
test morphology::benches::bench_grayscale_op_erode_small_image_point                  ... bench:       5,180 ns/iter (+/- 449)
```