@cospectrum 2024-05-09
> @cospectrum if you don't mind i would really appreciate your take on my last commit, i kinda struggled with making `disk` readable

The `master` version is much more beautiful.