@Morgane55440 2024-05-09
the [new commit](https://github.com/image-rs/imageproc/pull/598/commits/62eb25de885a898df0e329d713a589b1608fd155) is faster by about 20-30% for big images, but it is ***utterly unreadable***, i will rework it to make it even faster and significantly more readable and faster

also, i did not know about the minimum supported version, will rewrite the offending code