@cospectrum 2024-05-10
Write simple reference implementations and test against them