@theotherphil 2024-05-11
The `chunks` finding is interesting. This crate originally aimed to work on `GenericImage`s but I gave up on that when the performance overhead was too high. Now that we know we’re working with image buffers we can probably find a few places where existing functions can be sped up.

As this PR doesn’t change any APIs and comes with a significant speedup I’d like to get his merged for the next release - let me know when you’re ready for this to be merged. We can always simplify or clarify implementations later if necessary.