@cospectrum 2024-05-11
If it works, then I can simplify/refactor later myself, after the merge