@Morgane55440 2024-05-12
@theotherphil @cospectrum sorry for the wait was busy with life stuff. i finally got a decent optimization in, from 2x to almost 40x depending on the image and mask size/shape (the bigger the better the optimization)

i'm certain significant further improvements can be made, but i'm ok with shipping things as they are, except for the `lines` method which is a mess. i'm working on it. 

also the `apply` method is now unused. i think it's possible it might be useful later although it seems unlikely because it's just a very slow way to do things. should i remove it ?