@theotherphil 2024-05-12
Brilliant, thanks. Yes, remove `apply` if it’s not used.