@Morgane55440 2024-05-12
i have to say (working on `lines` amt), `[chunk_by](https://doc.rust-lang.org/nightly/core/primitive.slice.html#method.chunk_by)` is exactly what i need, that's really a shame. i'm on a good path to figuring out something good though